Using int64 to present graph execution latency.

+++ src/interface/graph.thrift
@@ -81,7 +81,7 @@ struct PlanDescription {
 
 struct ExecutionResponse {
     1: required common.ErrorCode        error_code;
-    2: required i32                     latency_in_us;  // Execution time on server
+    2: required i64                     latency_in_us;  // Execution time on server
     3: optional common.DataSet          data;
     4: optional binary                  space_name;
     5: optional binary                  error_msg;
